refactor(ev_msg): separate signed message submodule
